+++ src/composables/index.ts
@@ -10,6 +10,7 @@ export * from "./useLockState";
 export * from "./useNavDirection";
 export * from "./useOverlayBackHandler";
 export * from "./usePullToRefresh";
+export * from "./useScrollLock";
 export * from "./useSecretReveal";
 export * from "./useSecureScreen";
 export * from "./useSecuritySettings";